Connect up the Filecoin publisher
Allow the Filecoin publisher to be selected when choosing a publisher.
As we have to publish the data to IPFS as well as Filecoin, so that the
data can be retrieved by the client, this also adds a 'piggyback'
publisher that will call one publisher before attempting another and
return the results from the first.

This also:
* Moves the fallback publisher into a shared package with the
piggyback publisher
* Fixes a bug in `go_fmt.sh` when running with a deleted file
* Replaces uses of `ioutil.TempDir`/`os.MkdirTemp` with `t.TempDir` if
possible

Fixes #947